Updated README.md file
In this version, we have elaborated more on the users/personas of the project, as per the instructions provided in the feedback for the original project description draft. We have further added more details about storage and logging using Ceph's S3 API. Lastly, we have added our mentors' names and their email IDs as well as our names and email IDs in this draft.